src/config/__init__.py
# Copyright (c) 2025 Bytedance Ltd. and/or its affiliates
# SPDX-License-Identifier: MIT

# 导入环境变量配置
from dotenv import load_dotenv
load_dotenv() # 确保只加载一次

# 导入其他配置

from .agent_config import *
from .tools_config import *
from .yaml_loader import *

