pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auto-agents"
version = "0.1.0"
description = "Auto Agents for Task Automation"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "yuezhu", email = "your.email@example.com" },
]
dependencies = [
    "pip>=25.1.1",
    "duckduckgo-search>=8.0.2",
    "fastapi>=0.115.12",
    "graphviz>=0.20.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core==0.3.63",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.19",
    "langchain-tavily==0.2.0",
    "langgraph>=0.4.8",
    "markdownify>=1.1.0",
    "pandas==2.2.3",
    "python-dotenv>=1.1.0",
    "readabilipy>=0.3.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.3",
    "selenium>=4.33.0",
    "aiohttp==3.12.8",
    "greenlet==3.2.2",
    "langsmith==0.3.44",
    "webdriver-manager>=4.0.2",
    "pytest",
    "langgraph-supervisor",
    "ipython>=9.3.0",
    "langchain-mcp-adapters>=0.1.7",
    "mcp>=1.9.4",
]

[project.urls]
Documentation = "https://github.com/username/auto-agents#readme"
Source = "https://github.com/username/auto-agents"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    # 测试工具
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",  # 用于测试覆盖率报告
    
    # 代码质量工具
    "black>=24.1.0",
    "isort>=5.12.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",  # 类型检查
]

[project.scripts]
auto-agents = "src.main:main"
